ref: update linkableapi

+++ src/com/troblecodings/linkableapi/Linkingtool.java
@@ -25,9 +25,11 @@ import net.minecraft.world.World;
 
 public class Linkingtool extends Item {
 
-    private final BiPredicate<World, BlockPos> predicate;
-    private final Predicate<TileEntity> predicateSet;
-    private final TaggableFunction tagFromFunction;
+    protected static final String LINKINGTOOL_TAG = "linkingToolTag";
+
+    protected final BiPredicate<World, BlockPos> predicate;
+    protected final Predicate<TileEntity> predicateSet;
+    protected final TaggableFunction tagFromFunction;
 
     public Linkingtool(final ItemGroup tab, final BiPredicate<World, BlockPos> predicate) {
         this(tab, predicate, _u -> true);
@@ -49,70 +51,91 @@ public class Linkingtool extends Item {
 
     @Override
     public ActionResultType onItemUseFirst(final ItemStack stack, final ItemUseContext ctx) {
-        final World levelIn = ctx.getLevel();
         final PlayerEntity player = ctx.getPlayer();
         if (player == null)
             return ActionResultType.FAIL;
-        final BlockPos pos = ctx.getClickedPos();
+        final World levelIn = ctx.getLevel();
         if (levelIn.isClientSide)
             return ActionResultType.PASS;
+        final BlockPos pos = ctx.getClickedPos();
         final TileEntity entity = levelIn.getBlockEntity(pos);
+        final CompoundNBT itemTag = getOrCreateForStack(stack);
+        final CompoundNBT toolTag = itemTag.getCompound(LINKINGTOOL_TAG);
         if (entity instanceof ILinkableTile && this.predicateSet.apply(entity)) {
             final ILinkableTile controller = (ILinkableTile) entity;
             if (!player.isSneaking()) {
-                final CompoundNBT comp = stack.getTag();
-                if (comp == null) {
+                if (toolTag == null) {
                     message(player, "lt.notset", pos.toString());
                     return ActionResultType.PASS;
                 }
-                final BlockPos linkedPos = NBTUtil.readBlockPos(comp);
-                if (controller.link(linkedPos, comp)) {
+                final BlockPos linkedPos = NBTUtil.readBlockPos(toolTag);
+                if (controller.link(linkedPos, toolTag)) {
                     message(player, "lt.linkedpos", pos.getX(), pos.getY(), pos.getZ());
-                    stack.setTag(null);
+                    removeToolTag(stack);
                     message(player, "lt.reset");
+                    stack.hurtAndBreak(1, player,
+                            (user) -> user.broadcastBreakEvent(ctx.getHand()));
                     return ActionResultType.FAIL;
                 }
                 message(player, "lt.notlinked");
                 message(player, "lt.notlinked.msg");
                 return ActionResultType.FAIL;
             } else {
+                if (controller.canBeLinked() && predicate.test(levelIn, pos)) {
+                    final boolean containsPos =
+                            toolTag.contains("X") && toolTag.contains("Y") && toolTag.contains("Z");
+                    if (containsPos) {
+                        message(player, "lt.setpos.msg");
+                        return ActionResultType.FAIL;
+                    }
+                    final CompoundNBT newToolTag = NBTUtil.writeBlockPos(pos);
+                    tagFromFunction.test(levelIn, pos, newToolTag);
+                    itemTag.put(LINKINGTOOL_TAG, toolTag);
+                    message(player, "lt.setpos", pos.getX(), pos.getY(), pos.getZ());
+                    message(player, "lt.setpos.msg");
+                    return ActionResultType.SUCCESS;
+                }
                 if (controller.hasLink() && controller.unlink()) {
                     message(player, "lt.unlink");
+                    return ActionResultType.SUCCESS;
                 }
             }
             return ActionResultType.SUCCESS;
         } else if (predicate.test(levelIn, pos)) {
-            final CompoundNBT tag = stack.getTag();
-            if (tag != null) {
-                final boolean containsPos = tag.contains("X") && tag.contains("Y")
-                        && tag.contains("Z");
-                if (containsPos) {
-                    message(player, "lt.setpos.msg");
-                    return ActionResultType.FAIL;
-                }
+            final boolean containsPos =
+                    toolTag.contains("X") && toolTag.contains("Y") && toolTag.contains("Z");
+            if (containsPos) {
+                message(player, "lt.setpos.msg");
+                return ActionResultType.FAIL;
             }
-            final CompoundNBT comp = NBTUtil.writeBlockPos(pos);
-            tagFromFunction.test(levelIn, pos, comp);
-            stack.setTag(comp);
+            final CompoundNBT newToolTag = NBTUtil.writeBlockPos(pos);
+            tagFromFunction.test(levelIn, pos, newToolTag);
+            itemTag.put(LINKINGTOOL_TAG, newToolTag);
             message(player, "lt.setpos", pos.getX(), pos.getY(), pos.getZ());
             message(player, "lt.setpos.msg");
             return ActionResultType.SUCCESS;
-        } else if (player.isSneaking() && stack.getTag() != null) {
-            stack.setTag(null);
+        } else if (player.isSneaking()) {
+            removeToolTag(stack);
             message(player, "lt.reset");
             return ActionResultType.SUCCESS;
         }
         return ActionResultType.FAIL;
     }
 
+    public void removeToolTag(final ItemStack stack) {
+        getOrCreateForStack(stack).remove(LINKINGTOOL_TAG);
+    }
+
     @Override
     public void appendHoverText(final ItemStack stack, @Nullable final World levelIn,
             final List<ITextComponent> tooltip, final ITooltipFlag flagIn) {
-        final CompoundNBT tag = stack.getTag();
-        if (tag != null) {
-            final boolean containsPos = tag.contains("X") && tag.contains("Y") && tag.contains("Z");
+        final CompoundNBT nbt = getOrCreateForStack(stack);
+        if (nbt.contains(LINKINGTOOL_TAG)) {
+            final CompoundNBT comp = nbt.getCompound(LINKINGTOOL_TAG);
+            final boolean containsPos =
+                    comp.contains("X") && comp.contains("Y") && comp.contains("Z");
             if (containsPos) {
-                final BlockPos pos = NBTUtil.readBlockPos(tag);
+                final BlockPos pos = NBTUtil.readBlockPos(nbt);
                 tooltip(tooltip, "lt.linkedpos", pos.getX(), pos.getY(), pos.getZ());
                 return;
             }
@@ -121,10 +144,16 @@ public class Linkingtool extends Item {
         tooltip(tooltip, "lt.notlinked.msg");
     }
 
-    @SuppressWarnings({
-            "rawtypes", "unchecked"
-    })
-    public void tooltip(final List list, final String text, final Object... obj) {
+    public static CompoundNBT getOrCreateForStack(final ItemStack stack) {
+        CompoundNBT tag = stack.getTag();
+        if (tag == null) {
+            tag = new CompoundNBT();
+            stack.setTag(tag);
+        }
+        return tag;
+    }
+
+    public void tooltip(final List<ITextComponent> list, final String text, final Object... obj) {
         list.add(getComponent(text, obj));
     }
